Functionality for tracing keys or a set of keys using a regex.
I added a TraceSwitch to the Couchclient class and a KeyFilter which allows a regex or a string literal representing a key

to be used as a filter for tracing keys. Since not all TraceEventTypes are supported, I added a TraceSwitchExtensions class

with extension methods to simplify/abstrac things.

I also updated each method in CouchbaseClient that takes a key as a parameter to have a TraceSwitch.TraceKey(key) call.

This is somewhat of a nieve approach, but a better strategy can be added later as the use-case is further refined.

Likeswise the message is somewhat generic in that I only include the name of the method that invoked the tracing aling with

the key itself; improvement opportunities definitly exist.

Tracing can be enabled in the system.diagnostics section in the app.config and disabled by either removing the section or

changing the switch value to Off. Note that only Verbose, Information and Off are supported. Also, DefaultTraceListener and

ConsoleTraceListener are currently configured, but TextWriterTraceListener can be added to write the trace messages to

disk.A future task might be to integrate this into the CouchbaseClientConfiguration section.